+++ sdk/io/asn.1/basic/semantic/types.hpp
@@ -24,6 +24,7 @@ std::string nameof(const asn1_object* object);
 
 bool is_kind_of(const asn1_object* object, asn1_entity_t entity);
 
+bool is_kind_of_boolean(const asn1_object* object);
 bool is_kind_of_integer(const asn1_object* object);
 bool is_kind_of_real(const asn1_object* object);
 bool is_kind_of_cstring(const asn1_object* object);
@@ -31,6 +32,7 @@ bool is_kind_of_bstring(const asn1_object* object);
 bool is_kind_of_container(const asn1_object* object);
 bool is_kind_of_container_of(const asn1_object* object);
 
+bool is_kind_of_boolean(asn1_entity_t entity);
 bool is_kind_of_integer(asn1_entity_t entity);
 bool is_kind_of_real(asn1_entity_t entity);
 bool is_kind_of_cstring(asn1_entity_t entity);