+++ CHANGELOG.md
@@ -35,7 +35,8 @@
     * CBOR
 
 * details
-  * Revision 1080
+  * Revision 1081
+    * asn1_publisher - Constraints
   * Revision 1079
     * [tested] gcc 4.8.5
   * Revision 1078